instantiate a global tmpdir
We need a temporary directory that will get cleaned up only after the
session ends.  This is for the show command, which needs to be able to
load the image into an external viewer.  If we used a context manager
the file would get cleaned up before the viewer launched, and to avoid
that we would need some kind of artificial pause.